pyproject.toml
[tool.poetry]
name = "virbpy"
version = "0.1.0"
description = ""
authors = ["S. Co1 <sco1.git@gmail.com>"]
packages = [{include = "src"},]

[tool.poetry.scripts]
virbpy = "src.VirbPy:virbpy_cli"

[tool.poetry.dependencies]
python = "^3.8"
openpyxl = "^3.0"
pandas = "^1.2"
python-dateutil = "^2.8"
typer = "^0.3"

[tool.poetry.dev-dependencies]
black = "^21.5b1"
flake8 = "^3.8"
flake8-annotations = "^2.6"
flake8-bugbear = "^21.4"
flake8-docstrings = "^1.5"
flake8-fixme = "^1.1"
flake8-import-order = "^0.18"
flake8-tidy-imports = "^4.1"
mypy = "^0.812"
pep8-naming = "^0.11"
pre-commit = "^2.7"

[tool.black]
line-length = 100

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
